requirements.txt
absl-py                 1.4.0
accelerate              0.32.1
antlr4-python3-runtime  4.9.3
apex                    0.1
astunparse              1.6.3
attrs                   23.1.0
azure-core              1.28.0
azure-storage-blob      12.17.0
bcrypt                  4.0.1
beautifulsoup4          4.12.2
boto3                   1.26.118
botocore                1.29.118
Brotli                  1.0.9
cachetools              5.3.0
certifi                 2022.12.7
cffi                    1.15.1
charset-normalizer      3.1.0
circuitbreaker          1.4.0
colorlog                6.7.0
contourpy               1.1.0
cryptography            41.0.2
cycler                  0.11.0
decord                  0.6.0
diffusers==0.18.2
einops                  0.6.1
exceptiongroup          1.1.1
expecttest              0.1.4
filelock                3.12.0
fonttools               4.41.0
fsspec                  2024.6.1
ftfy                    6.1.1
fvcore                  0.1.5.post20221221
google                  3.0.0
google-auth             2.17.3
google-auth-oauthlib    1.0.0
grpcio                  1.54.0
huggingface-hub         0.23.4
hydra-colorlog          1.2.0
hydra-core              1.3.2
hypothesis              6.72.1
idna                    3.4
importlib-metadata      6.6.0
importlib-resources     6.0.0
iopath                  0.1.10
isodate                 0.6.1
Jinja2                  3.1.2
jmespath                1.0.1
kiwisolver              1.4.4
kornia                  0.6.12
lightning-utilities     0.9.0
Markdown                3.4.3
MarkupSafe              2.1.2
matplotlib              3.7.2
mosaicml-streaming      0.5.1
mpmath                  1.3.0
mypy-extensions         1.0.0
networkx                3.1
ninja==1.11.1
numpy                   1.24.3
omegaconf==2.3.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
peft==0.11.1
Pillow==9.5.0
pytorch3d==0.7.4
PyYAML==6.0
safetensors==0.4.3
scipy==1.10.1
timm==0.6.13
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
typing-inspect==0.9.0

